Removed dupilcated code lines, fix in import
